Add parsing of certain Yosys comments as code
This patch extends the reader layer of `#lang yosys` to parse certain
comments as code. This way, the information is preserved rather than
being discarded by the reader, so the expander can provide some useful
interpretation of it.

This patch extends the reader to preserve the following at the
top-level:

- yosys-smt2-stdt
- yosys-smt2-module
- yosys-smt2-topmod
- yosys-smt2-clock
- yosys-smt2-input
- yosys-smt2-output
- yosys-smt2-register
- yosys-smt2-memory

Preserving this would allow, e.g., collecting all the inputs into a
single data structure, so it would be possible to enumerate them with a
`(all-inputs)` without having to manually give a list of them elsewhere
in code that imports the Yosys module.

This patch also extends the reader to preserve comments in
`declare-datatype` forms, so that these human-readable names like
`cpu.cpu_state` may be exposed/used rather than the internal names like
`soc#6`.

This patch does not actually add any interpretations for any of these
comments, so it does not have any consequences for consumers of `#lang
yosys`. The `yosys-smt2-*` forms are expanded to `(void)`, and
similarly, the comments in `declare-datatype` forms are discarded.

Useful interpretations can be added in a later patch.