Switch struct to use human-readable field names
This patch follows up on 5f08089a8f4bfb104dbb70b6198e5aba1d101e89. It
switches the struct synthesized from the datatype declaration to use
human-readable names instead of the "modname#0" style of names. This is
a backwards-incompatible change, but it should be easy to adapt code to
use the new naming style.

Also, the old code used to export both the getters "modname_s-modname#0"
as well as wrappers for the getters with shorter names "modname#0". This
patch no longer exports the names in the latter style. Now, only the
getters are exported, and in the new style of "modname_s-fieldname".