Add exporting inputs/outputs/registers/memories
This patch adds functionality to export the set of inputs, outputs,
registers, and memories, where each element is a function taking in the
state. These sets are populated using the "yosys-smt2-input" style
comments that are present in the Yosys output.

Some of the work involved in producing these sets is done dynamically,
i.e. at run-time, because that seemed like the easiest/fastest way to
implement the functionality. It would be nicer to do it statically, i.e.
at compile-time. One way to do this would be to write a fancy
`#%module-begin` that processes the entire body and does the appropriate
thing, but this is probably not worth the effort, because the current
implementation of this functionality doesn't have any run-time cost
after startup, and the cost should be minimal.