Improve names of symbolic variables
We passed names to `define-symbolic*` to get good names for string
representations of symbolic variables. Now that we are extracting
human-readable names instead of the "module#0" style names, we should
use those human-readable as arguments to `define-symbolic*` instead of
the auto-generated numeric names.

This patch also makes a couple other small changes.
- Factor out the `(let () ...)` pattern into a new `fresh-symbolic`
  macro.
- Avoid passing the unused `name-stx` to the `zero-ctor` attribute,
  binding a syntax object directly (instead of a lambda that returns a
  syntax object). This also allows us to switch to `#:with` instead of
  `#:attr`.