Speed up compilation of large `(and ...)` terms
Rosette's `(and ...)` seems to have superlinear compilation time in terms of
the number of forms in the body.

Given a file of the following structure, with n forms inside the `and`:

    #lang rosette/safe
    (and #t #t ... #t)

The time to `raco make` grows quadratically:

  n  | time (s)
-----+---------
   0 |   0.696
 500 |   3.419
1000 |  12.083
1500 |  26.948
2000 |  47.020

This patch redefines `and` to compile down to a tree of Rosette `and` forms
with logarithmic depth, so no `and` has more than a constant number of terms
(currently 2). With this patch, the above example (with `#lang yosys`) runs
much faster:

   n  | time (s)
------+---------
    0 |  0.802
  500 |  1.039
 1000 |  1.227
 1500 |  1.472
 2000 |  1.754
 2500 |  1.961
 3000 |  2.103
 3500 |  2.266
 4000 |  2.495
 4500 |  2.799
 5000 |  3.198
 5500 |  3.427
 6000 |  3.605
 6500 |  3.902
 7000 |  4.030
 7500 |  4.302
 8000 |  4.474
 8500 |  4.751
 9000 |  5.007
 9500 |  5.140
10000 |  5.291