Change memoization strategy for memory/CPU savings
Previously, we relied on `define/memoize` from the memo package to
provide memoization for functions defined in code extracted by Yosys.
This memoization is critical for performance, because within a single
step, many of these functions are called multiple times with the same
arguments, and it is expensive to recompute them from scratch for every
call. Unfortunately, using `define/memoize` has an unfortunate side
effect of consuming too much memory.

The extracted functions all have a structure where they take a single
argument, the current state of the module, and produce some result.
`define/memoize` does memoization based on `eq?`, so memoized results
are only useful within the dynamic extent of a single call of the
transition function, because object identity of the state changes at
every step. But the memoized results are retained forever in the memo
table, so both the state (the key) and the result (the value) cannot be
garbage collected, which results in memory usage that grows linearly
over steps. In practice, this could result in over 100GB of extra memory
being used when running large examples for many steps. This also leads
to GC pressure and long pause times, slowing down execution.

This patch changes the memoization strategy, avoiding relying on the
memo package. We define a "memoization context", and we store
memoization information only within the dynamic extent of the context.
We create a fresh context for each call of the transition function,
because it's not useful to maintain memoization information across
calls. As for implementing the memoization itself, rather than having a
fully general strategy like `define/memoize`, we only remember the last
key/value that the function was called with. With the Yosys-extracted
code, within the extent of a single call of the transition function, all
the extracted functions should only ever be called with a single value
(the state), so it's unnecessary to do anything fancier like maintaining
a hash table for each memoized function.

This patch improves run-time performance, both in terms of CPU usage and
memory usage, and it has a nice side effect of improving compile time as
well. Computing the transition function is 3-5x faster, and memory use
no longer explodes. GC characteristics are much better, with the tail
latency decreased significantly. End-to-end performance, e.g. for
verifying deterministic start, is improved by over 5x in some cases.
Interestingly, compile times are also reduced by a factor of around 2x.