Optimize `distinct` and `=`
Yosys emits code where `distinct` is only called on bitvector values, so
we can use `bveq` instead of the more expensive `equal?`. Similarly, `=`
is almost always used on bitvector values (and rarely on booleans). We
could have implemented it as:

    (define (= x y) (if (boolean? x) (<=> x y) (bveq x y)))

Given that `=` is only rarely called on boolean values, we can avoid a
bit of overhead by writing a macro that implements the special behavior
for the rare uses on booleans, and otherwise compiles to `bveq`.

This improves performance of computing the step function by 20% on one
benchmark.