Fix short-circuiting behavior of `ite`
Previously, `ite` was defined as a function. This yielded correct
behavior: Yosys's extracted code is purely functional, so it's okay to
compute the value of the true and false branches and then return the
appropriate one based on the condition. However, this results in extra
computation, especially when the condition is a run-time constant
rather than a symbolic value, where we could have avoided computing the
value of one of the branches altogether.

The fix is easy: define `ite` as a macro that expands to Rosette's `if`.

This patch speeds up execution by more than 2x.