Switch conventions for mixed Racket/Rosette code
Instead of 'r:', now we use '@' as a prefix for Rosette identifiers in
`#lang racket` programs, which makes the code look similar to how the
Rosette library. And instead of 'racket:', which was pretty verbose, we
now use '!' as a prefix for Racket identifiers in `#lang rosette`
programs. The '!' prefix indicates that usage of these identifiers
should be thought about carefully, because it's potentially dangerous to
mix un-lifted Racket code with Rosette code.