Speed up uses of print-filter
It turns out that Racket calls the `write-proc` twice for explicit calls
to `write` / `display` / `print`. It seems that this is to detect
cycles. We do not have cycles in this data structure, but it seems that
there is no easy way to turn off this behavior of Racket. This is
unfortunate, especially when using `print-filter`, because we pay the
cost of converting the entire struct to a string even if we only want to
print a single field.

This patch uses the `print-filter` in `write-proc` even when in
write/display mode, so that the second call is not too slow.