Tweak import paths for yosys
This patch makes `#lang yosys` code use `yosys/verilog` as the module
path, so that `yosys` is free to be used for other purposes. Now,
`yosys/main.rkt` can export all identifiers that are useful to require
in hand-written code that imports a `#lang yosys` module. Right now,
this only contains parameters, so a `(require yosys/parameters)` can now
be replaced with a `(require yosys)`. In the future, as the set of
useful functionality grows, it will be convenient to be able to import a
single module to get all relevant functionality.