Add generics for Yosys-related functionality
This makes it so we can avoid things like `struct->vector` and applying
the constructor to map over the fields of a Yosys-generated struct. As
another benefit, this lets us avoid having to pass around things like
the constructor.

This patch doesn't seem to impact performance. In a rough benchmark of
several examples (from outside this repo), the run times do not seem to
be significantly different before/after this patch (if anything, run
times are slightly better).