Make fresh-symbolic into a function
This patch makes `fresh-symbolic` into a function, which means that it
can generate new symbolic constants that have printed names that are
based on run-time values rather than syntax.

Old uses of fresh-symbolic can be converted as follows:

    (fresh-symbolic foo type)

Can be changed to:

    (fresh-symbolic 'foo type)

Additionally, because `fresh-symbolic` is now a function, things like
this are possible:

    (let ([x 'hello])
      (fresh-symbolic x boolean?))
    (print x)  ; prints "hello%0"

Whereas previously, this would have printed "x$0" (and it wouldn't have
been possible to name a symbolic constant after the run-time value
'hello.