Add support for reasoning about inputs and outputs
This seems to work pretty well, and the performance when symbolic
inputs/outputs aren't used basically matches the performance from
before. Symbolic inputs add some overhead, and hints are needed to
decide when to abstract terms. The symbolic terms also produce lots of
garbage, and it seems that calls to `(collect-garbage)` are helpful
(these can also be supplied as hints).