Facilitate garbage collection in Rosette code
Rosette uses a term cache that stores all symbolic constants and terms.
For a Rosette program that creates lots of garbage, this results in
increased memory pressure (because Rosette terms effectively can't be
GCed), which also translates to slowdown.

This is a simple program (`gc.rkt`) that demonstrates the issue:

    #lang rosette

    (define (silly)
      (define-symbolic* x y z integer?)
      (+ x y))

    (define (sillyn n)
      (cond
        [(<= n 0) (silly)]
        [else (silly) (sillyn (sub1 n))]))

    (sillyn 5000000)

    (collect-garbage)

    (printf "current memory use: ~a MB~n"
      (~r (/ (current-memory-use) (expt 2 20)) #:precision 1))

On my machine, running `raco make gc.rkt && time gc.rkt` produces:

    current memory use: 2835.4 MB
    racket gc.rkt  126.50s user 10.05s system 99% cpu 2:16.80 total

This patch introduces tools to replace the term cache with a structure
that enables garbage collection of otherwise unreachable objects, using
a weak hash with ephemerons for the values. The ephemerons are necessary
because the values in the hash table sometimes refer back to the keys.

With this patch and adding a call to `(term-cache-ephemeral!)`, both the
memory usage and run time is reduced for `gc.rkt`:

    current memory use: 44.5 MB
    racket gc.rkt  63.93s user 0.95s system 100% cpu 1:04.87 total