Work around struct-copy hygiene bug
`struct-copy` looks up fields by concatenating the struct name with the
field name. In our use, the struct name is written at the `update-name`
and `make-name` macro definition sites, and so if a yosys module is
imported with renaming (e.g. by `prefix-in`), there's a mismatch in the
field accessors. For example, in the test case added in this patch, with
the old code for `update-name`, `struct-copy` would try to look up
`counter_s-count` even though the accessor is bound to
`count:counter_s-count`, resulting in a compile-time error.

This bug in `struct-copy` was fixed in
https://github.com/racket/racket/pull/2700 and
https://github.com/racket/racket/pull/3267, but this is a recent fix, so
it's included in nightlies but not the latest release version of Racket
(v7.7).

Rather than require using a nightly, this patch works around the bug by
using the `update-fields` generic method.

This patch also fixes a bug in the default implementation of
`update-fields`. The implementation used `hash-ref` and used a
`failure-result` fallback argument (the previous value of the field).
This is either a value or a procedure, and if it's a procedure, it is
called. This works fine with booleans/bitvectors/vectors, but it doesn't
work with uninterpreted functions, because those are recognized as
procedures and called. We work around this issue by always using the
procedure version of `failure-result`, supplying a lambda that returns
the old value.