Unify field selection between print and concretize
We have a nice interface to print a subset of fields, being able to
select fields with a parameter that can be one of several forms, making
it easy to do substring match or regex match against fields. This patch
renames this parameter from `print-filter` to `field-filter` and makes
`concretize-fields` also use this same parameter for selecting a subset
of fields. This makes it easier to e.g. concretize all "uart.*" fields
without having to explicitly list them, which was necessary prior to
this patch.