Auto-gen input/output structs and expose metadata
This isn't quite compatible with all possible Yosys output formats:
sometimes Yosys does weird things to an input bus, and rather than
representing it as a single bitvector as part of the state, it splits it
up. The old "multixor" test had this problem. This doesn't seem to come
up in the real examples we are working with, so this patch ignores this
issue for now. If a module we work with in the future is incompatible
with this change, it will crash when compiling (it won't silently give
buggy results), which seems fine.